Rename existingThoughtDelete -> deleteThought
