+++ src/action-creators/index.ts
@@ -35,7 +35,7 @@ export const error = reducerToThunk<typeof reducers.error>('error')
 export const expandHoverTop = reducerToThunk<typeof reducers.expandHoverTop>('expandHoverTop')
 export const expandBottom = reducerToThunk<typeof reducers.expandBottom>('expandBottom')
 export const editThought = reducerToThunk<typeof reducers.editThought>('editThought')
-export const existingThoughtDelete = reducerToThunk<typeof reducers.existingThoughtDelete>('existingThoughtDelete')
+export const deleteThought = reducerToThunk<typeof reducers.deleteThought>('deleteThought')
 export const moveThought = reducerToThunk<typeof reducers.moveThought>('moveThought')
 export const extractThought = reducerToThunk<typeof reducers.extractThought>('extractThought')
 export const indent = reducerToThunk<typeof reducers.indent>('indent')
