+++ src/reducers/index.ts
@@ -29,7 +29,7 @@ export { default as error } from './error'
 export { default as expandHoverTop } from './expandHoverTop'
 export { default as expandBottom } from './expandBottom'
 export { default as editThought } from './editThought'
-export { default as existingThoughtDelete } from './existingThoughtDelete'
+export { default as deleteThought } from './deleteThought'
 export { default as moveThought } from './moveThought'
 export { default as expandContextThought } from './expandContextThought'
 export { default as extractThought } from './extractThought'
